+++ Core/Src/robot_control.c
@@ -23,8 +23,8 @@ void setLocalTargetSpeed(RobotCommandV2 * ai_cmd, target_t * target, imu_t * imu
 
   } else if (ai_cmd->control_mode == POLAR_VELOCITY_TARGET_MODE) {
     float temp_gloval_vel[2];
-    temp_gloval_vel[0] = ai_cmd->mode_args.polar_velocity.target_global_velocity_r * cos(ai_cmd->mode_args.polar_velocity.target_global_velocity_theta);
-    temp_gloval_vel[1] = ai_cmd->mode_args.polar_velocity.target_global_velocity_r * sin(ai_cmd->mode_args.polar_velocity.target_global_velocity_theta);
+    temp_gloval_vel[0] = ai_cmd->mode_args.polar_velocity.target_global_velocity_r * cosf(ai_cmd->mode_args.polar_velocity.target_global_velocity_theta);
+    temp_gloval_vel[1] = ai_cmd->mode_args.polar_velocity.target_global_velocity_r * sinf(ai_cmd->mode_args.polar_velocity.target_global_velocity_theta);
 
     convertGlobalToLocal(temp_gloval_vel, target->local_vel, imu->yaw_rad);
